+++ bindings/python/src/connection.rs
@@ -3,7 +3,7 @@ use crate::*;
 use std::sync::Arc;
 use pyo3_async_runtimes::tokio::future_into_py;
 
-// Connection to a Fluss cluster
+/// Connection to a Fluss cluster
 #[pyclass]
 pub struct FlussConnection {
     inner: Arc<fcore::client::FlussConnection>,
@@ -11,20 +11,13 @@ pub struct FlussConnection {
 
 #[pymethods]
 impl FlussConnection {
-    // Create a new FlussConnection from a config
     #[new]
     fn new(config: &Config) -> PyResult<Self> {
-        // For now, we'll create a placeholder connection
-        // In a real implementation, we'd establish the connection here
-        let rust_config = config.get_core_config();
-        
-        // We'll need to use a different approach since we can't do async in __new__
-        // For now, create a mock connection
-        // This is a limitation - we might need to refactor this later
-        Err(FlussError::new_err("FlussConnection constructor not yet implemented. Use FlussConnection.connect() instead."))
+        // Always use connect to create a new connection
+        Err(FlussError::new_err("Use FlussConnection.connect() instead."))
     }
 
-    // Create a new FlussConnection (async)
+    /// Create a new FlussConnection (async)
     #[staticmethod]
     fn connect<'py>(py: Python<'py>, config: &Config) -> PyResult<Bound<'py, PyAny>> {
         let rust_config = config.get_core_config();
@@ -44,7 +37,7 @@ impl FlussConnection {
         })
     }
     
-    // Get admin interface
+    /// Get admin interface
     fn get_admin<'py>(&self, py: Python<'py>) -> PyResult<Bound<'py, PyAny>> {
         let client = self.inner.clone();
 
@@ -61,7 +54,7 @@ impl FlussConnection {
         })
     }
 
-    // Get a table
+    /// Get a table
     fn get_table<'py>(&self, py: Python<'py>, table_path: &TablePath) -> PyResult<Bound<'py, PyAny>> {
         let client = self.inner.clone();
         let core_path = table_path.to_core().clone();
